Theodw 1748 Reorganise tests (#1916)
* Add out-of-hours pipelines to validate the data

* Reorganise tests scripts. Update validation pipeline to run validation tests

* Fix syntax error

* Remove conditions

* Create separate conftest for each kind of test. Improve stlying:

* Add missing params

* Reorganise unit tests ado job

* Add integration test and e2e test folders

* Rename test folder

* Update smoke and integration test jobs to run corresponding pytest folders

* Rename file

* Add blank test to allow the CICD pipeline to pass when there are no other tests

* Fix broken import

* Add placeholder test

* Prevent Pull request triggering the pipeline
