+++ pipelines/jobs/run-synapse-data-validation-tests.yaml
@@ -3,11 +3,12 @@ parameters:
   env: ''
 
 ##
-# Run Synapse unit tests
+# Run data validation tests
 ##
 jobs:
-- job: RunSynapseUnitTestsJob
+- job: RunDataValidationTestsJob
   pool: ${{ parameters.agentPool }}
+  timeoutInMinutes: 0 # Max timeout
   steps:
   - checkout: self
     clean: true
@@ -21,5 +22,5 @@ jobs:
       export CREDENTIAL_NAME="${{ format('https://dev.azuresynapse.net/.default', lower(parameters.env)) }}"
       echo $SYNAPSE_ENDPOINT 
       echo $CREDENTIAL_NAME
-      python3 -m pytest $(Build.SourcesDirectory)/workspace/testing/unit_test -vv -rP -n 4
+      python3 -m pytest $(Build.SourcesDirectory)/workspace/testing/validation_test -vv -rP -n 4
     displayName: 'pytest'
