+++ pipelines/jobs/run-synapse-smoketests.yaml
@@ -11,11 +11,15 @@ jobs:
   steps:
   - checkout: self
     clean: true
-  - task: AzureCLI@2
-    displayName: 'Testing'
-    inputs:
-      azureSubscription: 'ODW ${{ upper(parameters.env) }} - Infrastructure'
-      scriptType: 'bash'
-      scriptLocation: 'inlineScript'
-      inlineScript: |
-        echo "No tests to run"
+  - template: ${{variables['System.DefaultWorkingDirectory']}}/pipelines/steps/azure-login.yaml
+  - script: |
+      export PATH="$PATH:/home/AzDevOps/.local/bin/"
+      sudo apt-get install unixodbc-dev
+      sudo curl -fsSL https://aka.ms/install-azd.sh | bash
+      pip install pandas pyarrow pytest pytest-azurepipelines aiohttp pyodbc azure-storage-blob azure-keyvault-secrets azure-identity azure-storage-file-datalake azure-synapse azure-mgmt-resource pytest pytest-xdist[psutil]
+      export SYNAPSE_ENDPOINT="${{ format('https://pins-synw-odw-{0}-uks.dev.azuresynapse.net/', lower(parameters.env)) }}"
+      export CREDENTIAL_NAME="${{ format('https://dev.azuresynapse.net/.default', lower(parameters.env)) }}"
+      echo $SYNAPSE_ENDPOINT 
+      echo $CREDENTIAL_NAME
+      python3 -m pytest $(Build.SourcesDirectory)/workspace/testing/smoke_test -vv -rP -n 4
+    displayName: 'pytest'
