+++ pipelines/nightly-data-validation-tests-dev.yaml
@@ -1,0 +1,26 @@
+name: Data Validation Dev
+
+# This is the nightly data validation pipeline against the dev environment
+
+# Automatically run against the dev branch at a set time every day
+trigger: none
+pr: none
+schedules:
+- cron: '0 20 * * 1-5' # cron syntax defining a schedule. At 8pm every week day
+  displayName: Daily dev branch run against the dev env
+  branches:
+    include:
+    - dev
+  always: true
+  batch: false
+
+variables:
+- name: env
+  value: dev
+- group: "Terraform Dev"
+
+stages:
+- template: stages/run-validation-tests-stage.yaml
+  parameters:
+    agentPool: 'pins-agent-pool-odw-${{ variables.env }}-uks'
+    env: ${{ variables.env }}
