+++ workspace/testing/integration_test/test_placeholder.py
@@ -1,0 +1,3 @@
+def test_placeholder():
+    # Remove this test when other integration tests have been created
+    pass
