+++ workspace/testing/smoke_test/test_placeholder.py
@@ -1,0 +1,3 @@
+def test_placeholder():
+    # Remove this test when other smoke tests have been created
+    pass
