+++ workspace/testing/validation_test/test_appeal_document.py
@@ -1,21 +1,20 @@
 import pytest
-import pipelineutils
-import constants
+import testing.util.constants as constants
+import workspace.testing.util.pipelineutils as pipelineutils
 import warnings
 
-def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoint: str):
+def test_appeal_document_notebook(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_nsip_project"
-    
+    notebookname: str = "py_unit_tests_appeal_document"
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "nsip-project",
+               "value": "appeal-document",
             },
             "std_db_name": {
                "type": "String",
@@ -31,35 +30,19 @@ def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoi
             },
             "std_table_name": {
                "type": "String",
-               "value": "sb_nsip_project",
+               "value": "sb_appeal_document",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "sb_nsip_project",
+               "value": "sb_appeal_document",
             },
             "hrm_table_final": {
                "type": "String",
-               "value": "nsip_project",
+               "value": "appeals_document_metadata",
             },
             "curated_table_name": {
                "type": "String",
-               "value": "nsip_project",
-            },
-            "curated_db_migration_name": {
-               "type": "String",
-               "value": "odw_curated_migration_db",
-            },
-            "curated_table_migration_name": {
-               "type": "String",
-               "value": "nsip_project",
-            },
-            "primary_key": {
-               "type": "String",
-               "value": "caseId",
-            },
-            "migration_primary_key": {
-               "type": "String",
-               "value": "caseReference",
+               "value": "appeal_document",
             }
         }
     }
@@ -70,7 +53,7 @@ def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoi
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_nsip_project Completed")
+    print("test_appeal_document Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
