+++ workspace/testing/validation_test/test_pins_lpa_curated.py
@@ -1,21 +1,21 @@
 import pytest
-import pipelineutils
-import constants
+import workspace.testing.util.pipelineutils as pipelineutils
+import testing.util.constants as constants
 import warnings
 
-def test_relevant_representation_notebook(credential_name, azure_credential, synapse_endpoint: str):
+def test_pins_lpa_curated(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_relevant_representation"
+    notebookname: str = "py_unit_tests_pins_lpa_curated"
     
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "nsip-representation",
+               "value": "pins-lpa",
             },
             "std_db_name": {
                "type": "String",
@@ -29,34 +29,18 @@ def test_relevant_representation_notebook(credential_name, azure_credential, syn
                "type": "String",
                "value": "odw_curated_db",
             },
-            "curated_migration_db_name": {
-               "type": "String",
-               "value": "odw_curated_migration_db",
-            },
             "std_table_name": {
                "type": "String",
-               "value": "sb_nsip_representation",
+               "value": "pins_lpa",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "sb_nsip_representation",
-            },
-            "hrm_table_final": {
-               "type": "String",
-               "value": "nsip_representation",
+               "value": "pins_lpa",
             },
             "curated_table_name": {
                "type": "String",
-               "value": "nsip_representation",
+               "value": "pins_lpa",
             },
-            "std_hzn_table_name": {
-               "type": "String",
-               "value": "horizon_nsip_relevant_representation",
-            },
-            "curated_migration_table_name": {
-               "type": "String",
-               "value": "nsip_representation",
-            }
         }
     }
     notebook_raw_params.update(constants.SPARK_POOL_CONFIG)
@@ -66,7 +50,7 @@ def test_relevant_representation_notebook(credential_name, azure_credential, syn
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_relevant_representation Completed")
+    print("test_pins_lpa_curated Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
